Add support to set/change log retention period (#4)
Issue #, if available:

Description of changes:
This change adds the support to set/change the retention period for log-groups. By default, when log group is created, log retention period is set to 'never expire'. So, it is necessary to allow to change it.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.